+++ lib/daru/io/importers/rails_log.rb
@@ -1,9 +1,11 @@
 require 'daru/io/importers/base'
+require 'daru/io/importers/shared/request_log_analyzer_patch'
 
 module Daru
   module IO
     module Importers
       class RailsLog < Base
+        include RequestLogAnalyzerPatch
         Daru::DataFrame.register_io_module :read_rails_log, self
         Daru::DataFrame.register_io_module :from_rails_log, self
 
@@ -14,18 +16,19 @@ module Daru
 
         def read(path)
           @path = path
-          @file_data = Base.parse_log(@path,:rails3)
+          @file_data = RequestLogAnalyzerPatch.parse_log(@path,:rails3)
           self
         end
 
-        def call()
-          ind = [:method, :path, :ip, :timestamp, :line_type, :lineno, :source,
+        INDEX = [:method, :path, :ip, :timestamp, :line_type, :lineno, :source,
                  :controller, :action, :format, :params, :rendered_file,
                  :partial_duration, :status, :duration, :view, :db]
-          data = Daru::DataFrame.new({},index: ind).transpose
+
+        def call()
+          data = Daru::DataFrame.new({},index: INDEX).transpose
           @file_data.each do |hash|
             row = []
-            ind.each do |attr|
+            INDEX.each do |attr|
               row << hash[attr]
             end
             data.add_row row
